+++ packages/hash/frontend/src/shared/layout/layout-with-header.tsx
@@ -1,0 +1,14 @@
+import { ReactNode, VFC } from "react";
+import { PageHeader } from "./layout-with-header/page-header";
+import { PlainLayout } from "./plain-layout";
+
+export const LayoutWithHeader: VFC<{
+  children?: ReactNode;
+}> = ({ children }) => {
+  return (
+    <PlainLayout>
+      <PageHeader />
+      {children}
+    </PlainLayout>
+  );
+};
