+++ packages/hash/frontend/src/shared/layout/layout-with-header/search-bar.tsx
