+++ packages/hash/frontend/src/shared/layout/layout-with-header/shared/header-icon-button.tsx
@@ -1,5 +1,5 @@
 import { styled } from "@mui/material";
-import { IconButton } from "../../ui";
+import { IconButton } from "../../../ui";
 
 const focusBorderOffset = 6;
 const focusBorderWidth = 3;
