+++ packages/hash/frontend/src/shared/layout/layout-with-sidebar/account-entity-type-list.tsx
