+++ packages/hash/frontend/src/shared/layout/layout-with-sidebar/account-entity-type-list/entity-type-menu.tsx
