+++ packages/hash/frontend/src/shared/layout/layout-with-sidebar/page-sidebar.tsx
@@ -1,0 +1,2 @@
+export * from "./sidebar";
+export * from "./sidebar-context";
