+++ packages/hash/frontend/src/shared/layout/layout-with-sidebar/sidebar-context.tsx
