+++ packages/hash/frontend/src/shared/layout/layout-with-sidebar/workspace-switcher.tsx
