--- packages/hash/frontend/src/shared/layout/page-layout.tsx
@@ -1,4 +1,0 @@
-export * from "./page-layout/layout-with-sidebar";
-export * from "./page-layout/plain-layout";
-export * from "./page-layout/plain-layout-with-header";
-export * from "./page-layout/types";
