--- packages/hash/frontend/src/shared/layout/page-layout/types.ts
@@ -1,6 +1,0 @@
-import { NextPage } from "next";
-import { ReactElement, ReactNode } from "react";
-
-export type NextPageWithLayout<T = {}> = NextPage<T> & {
-  getLayout?: (page: ReactElement) => ReactNode;
-};
