--- packages/hash/frontend/src/shared/layout/page-sidebar.tsx
@@ -1,2 +1,0 @@
-export * from "./page-sidebar/sidebar";
-export * from "./page-sidebar/sidebar-context";
